+++ crates/alkanes-runtime/src/stdio.rs
@@ -8,7 +8,7 @@ impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> Result<(), Error> {
         let mut data = to_arraybuffer_layout::<Vec<u8>>(s.to_string().as_bytes().to_vec());
         unsafe {
-            __log(to_passback_ptr(&mut data));
+            __log(to_passback_ptr(&mut data) as i32);
         }
         return Ok(());
     }
